refactor(project-files-view): extract query model owner
