+++ src/renderer/src/stores/session-store.test.ts
@@ -2786,7 +2786,6 @@ describe('session store public contract', () => {
       'src/renderer/src/pages/workspace/DeleteSessionDialog.tsx',
       'src/renderer/src/pages/workspace/DownloadSessionArtifactsDialog.tsx',
       'src/renderer/src/pages/workspace/PreviewFileSurface.tsx',
-      'src/renderer/src/pages/workspace/ProjectFilesView.tsx',
       'src/renderer/src/pages/workspace/RenameSessionDialog.tsx',
       'src/renderer/src/pages/workspace/SessionNotebookDialog.tsx',
       'src/renderer/src/pages/workspace/WorkspaceActivityIcon.tsx',
@@ -2808,6 +2807,7 @@ describe('session store public contract', () => {
       'src/renderer/src/pages/workspace/preview-file-item.ts',
       'src/renderer/src/pages/workspace/previews/PreviewToolContent.tsx',
       'src/renderer/src/pages/workspace/project-files-library.ts',
+      'src/renderer/src/pages/workspace/project-files-query-model.ts',
       'src/renderer/src/pages/workspace/session-plan/active-branch-plan.ts',
       'src/renderer/src/pages/workspace/session-plan/respond-to-session-plan.ts',
       'src/renderer/src/pages/workspace/use-project-artifact-files.ts',
